refactor: rename package to charmed-kubeflow-chisme
Merge after #15
