+++ setup.cfg
@@ -1,14 +1,11 @@
 [metadata]
-name = k8s-resource-handler
+name = charmed-kubeflow-chisme
 version = 0.0.1
 author = Charmed Kubeflow
-;author_email = author@example.com
 description = A collection of helpers for Charms maintained by the Charmed Kubeflow team
 long_description = file: README.md
 long_description_content_type = text/markdown
-url = https://github.com/canonical/k8s-resource-handler
-project_urls =
-    Bug Tracker = https://github.com/canonical/k8s-resource-handler/issues
+url = https://github.com/canonical/charmed-kubeflow-chisme
 classifiers =
     Programming Language :: Python :: 3
     License :: OSI Approved :: MIT License
