+++ src/charmed_kubeflow_chisme/lightkube/batch/_many.py
