+++ src/charmed_kubeflow_chisme/status_handling/_get_first_worst_error.py
